requirements.txt
bitsandbytes==0.44.1
bitarray==3.0.0
datasets==3.0.2
einops==0.8.1
memory-profiler==0.61.0
transformers==4.39.3
wandb==0.19.11
python-dotenv==1.0.1
python-logging-loki==0.3.1
speedtest-cli==2.1.3
loguru==0.7.2
flake8==7.0.0
black==23.7.0
isort==5.13.2
expecttest==0.2.1
torch==2.5.1
influxdb-client==1.48.0
sentencepiece==0.2.0
openskill==6.1.3
muon-optimizer @ git+https://github.com/KellerJordan/Muon@f90a42b28e00b8d9d2d05865fe90d9f39abcbcbd
rich==14.1.0
bittensor-cli==9.11.2
boto3==1.40.45
